+++ lib/base/base_controller.dart
@@ -20,9 +20,19 @@ abstract class LifecycleControl {
   void onPause() {}
 }
 
+abstract class KeyboardControl {
+  /// 键盘弹出
+  void onKeyboardShow() {}
+
+  /// 键盘收起
+  void onKeyboardHide() {}
+}
+
 abstract class BaseController<T extends BaseModel> extends GetxController
-    with WidgetsBindingObserver, LifecycleControl
+    with WidgetsBindingObserver, LifecycleControl, KeyboardControl
     implements ModelSetter<T>, ActionControl {
+  late BuildContext context;
+
   /// 是否可以返回，用于拦截返回按钮操作
   @override
   Future<bool> onWillPop() {
@@ -48,6 +58,19 @@ abstract class BaseController<T extends BaseModel> extends GetxController
     } else if (state == AppLifecycleState.paused) {
       onPause();
     }
+  }
+
+  @override
+  void didChangeMetrics() {
+    super.didChangeMetrics();
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      bool hide = MediaQuery.of(context).viewInsets.bottom == 0;
+      if (hide) {
+        onKeyboardHide();
+      } else {
+        onKeyboardShow();
+      }
+    });
   }
 }